# Use new formulas from Jorge Zuniga to compute Catalan and zeta(3) faster