… doc/source/references.rst