See #1622.

This speeds up Catalan's constant 2x (7.6s -> 3.9s for 10M digits) and zeta(3) 1.5x (3.2s -> 2.1s for 10M digits).